requirements.txt
torch==2.8.0
pandas==2.3.3
pillow==11.3.0
torchvision==0.23.0